Removed calls from compute_joint_norm_totals_opt to norm_total property
The compute_joint_norm_totals_opt function in topo/sheet/optimized.py
was accessing the Python property norm_total for each CF, which is
a) a lot slower and b) in previous versions entailed 32-bit accuracy
summation of weights (something that has since been fixed). This commit
does the summation in C, speeding up the operation and making it
more consistent with other optimized functions.