version: Fixed the CalledProcessError problem
This was due to CalledProcessError not being a built-in error, causing
`raise CalledProcessError` to think the error to be raised was in a
variable called `CalledProcessError`.

I could not reproduce this issue on either Python 2.6.8 nor 2.7.3, but I
did reproduce it on 2.6.6 and there are reports of people encountering this
on "some" version of 2.7.