Fixed GUI bug involving broken measurements
When running training, stopping it before the full number of specified
steps was up and then performing a measurement, all through the GUI would
result in the ongoing_measurement flag to be erroneously set to false,
which would interrupt the measurement. This fix ensures that when
iniating a measurement the timer stop flag is set to False.