Reverted to storing timestamps as mpq
Previously comparisons between mpqs was broken on some systems
(e.g. DICE), timestamps were therefore cast to floats. This is unsafe
therefore we are reverting. There was a bug in gmpy that caused
some comparisons to fail, reversing the order of the comparison should
address this problem. If error recurs there needs to be an alternative
fix or a requirement for the latest version of gmpy.