sys.path initialization moved to external/__init__.py
Initialization of the sys.path with topographica and the external
submodules is now performed by external/__init__.py. This
mechanism is used by the topographica script, the IPython
Notebook profile and nosetests via the -t unit and -t coverage
options. Nosetests can only be run directly if all the submodules
can be found on the PYTHONPATH.