Updated GCAL notebook to use new dataviews system
Note that the results have changed from those shown in the previous
version (both in the notebook and in the static HTML version). This is
due to a change to the default input seed (from 42 to 102) that was
made without a corresponding update to the notebook contents. From now
on, notebooks should only be committed if their outputs have been
cleared first.